refactor: URL 관련 함수 이동
