+++ src/app/providers/PostProvider.tsx
@@ -1,26 +1,12 @@
+import { fetchPostsAPI } from "@/entities/post/api/fetch-posts";
 import { Post, Tag } from "@/entities/post/models/post.types";
 import { User } from "@/entities/user/models/user.types";
 import { PostContext } from "@/features/posts/models/use-post-context";
+import { useURLContext } from "@/features/posts/models/use-url-context";
 import { ReactNode, useEffect, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
 
-type PostsData = {
-  posts: Post[];
-  total: number;
-  skip: number;
-  limit: number;
-};
-
-type UsersData = {
-  users: User[];
-  total: number;
-  skip: number;
-  limit: number;
-};
 export const PostProvider = ({ children }: { children: ReactNode }) => {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const queryParams = new URLSearchParams(location.search);
+  const { limit, skip, sortBy, sortOrder, selectedTag, updateURL } = useURLContext();
   const [loading, setLoading] = useState(false);
 
   const [posts, setPosts] = useState<Post[]>([]);
@@ -28,51 +14,25 @@ export const PostProvider = ({ children }: { children: ReactNode }) => {
 
   // search
   const [total, setTotal] = useState(0);
-  const [skip, setSkip] = useState(parseInt(queryParams.get("skip") || "0"));
-  const [limit, setLimit] = useState(parseInt(queryParams.get("limit") || "10"));
-  const [searchQuery, setSearchQuery] = useState(queryParams.get("search") || "");
-
-  // sort
-  const [sortBy, setSortBy] = useState(queryParams.get("sortBy") || "");
-  const [sortOrder, setSortOrder] = useState(queryParams.get("sortOrder") || "asc");
 
   // 게시물 가져오기
-  const fetchPosts = () => {
+  const fetchSetPosts = async () => {
     setLoading(true);
-    let postsData: PostsData;
-    let usersData: User[];
-
-    fetch(`/api/posts?limit=${limit}&skip=${skip}`)
-      .then((response) => response.json())
-      .then((data) => {
-        postsData = data;
-        return fetch("/api/users?limit=0&select=username,image");
-      })
-      .then((response) => response.json())
-      .then((users: UsersData) => {
-        usersData = users.users;
-
-        const postsWithUsers: Post[] = postsData.posts.map((post) => {
-          const author = usersData.find((user) => user.id === post.userId);
-          if (!author) throw new Error("wrong author");
-          return { ...post, author };
-        });
-
-        setPosts(postsWithUsers);
-        setTotal(postsData.total);
-      })
-      .catch((error) => {
-        console.error("게시물 가져오기 오류:", error);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
+    try {
+      const { posts, total } = await fetchPostsAPI(limit, skip);
+      setPosts(posts);
+      setTotal(total);
+    } catch (error) {
+      console.error("게시물 조회 오류", error);
+    } finally {
+      setLoading(false);
+    }
   };
 
   // 태그별 게시물 가져오기
   const fetchPostsByTag = async (tag: Post["tags"][number]) => {
     if (!tag || tag === "all") {
-      fetchPosts();
+      fetchSetPosts();
       return;
     }
     setLoading(true);
@@ -98,21 +58,8 @@ export const PostProvider = ({ children }: { children: ReactNode }) => {
     setLoading(false);
   };
 
-  // 게시물 삭제
-  const deletePost = async (id: Post["id"]) => {
-    try {
-      await fetch(`/api/posts/${id}`, {
-        method: "DELETE",
-      });
-      setPosts(posts.filter((post) => post.id !== id));
-    } catch (error) {
-      console.error("게시물 삭제 오류:", error);
-    }
-  };
-
   // tag
   const [tags, setTags] = useState<Tag[]>([]);
-  const [selectedTag, setSelectedTag] = useState(queryParams.get("tag") || "");
 
   // 태그 가져오기
   const fetchTags = async () => {
@@ -125,18 +72,6 @@ export const PostProvider = ({ children }: { children: ReactNode }) => {
     }
   };
 
-  // URL 업데이트 함수
-  const updateURL = () => {
-    const params = new URLSearchParams();
-    if (skip) params.set("skip", skip.toString());
-    if (limit) params.set("limit", limit.toString());
-    if (searchQuery) params.set("search", searchQuery);
-    if (sortBy) params.set("sortBy", sortBy);
-    if (sortOrder) params.set("sortOrder", sortOrder);
-    if (selectedTag) params.set("tag", selectedTag);
-    navigate(`?${params.toString()}`);
-  };
-
   useEffect(() => {
     fetchTags();
   }, []);
@@ -145,21 +80,11 @@ export const PostProvider = ({ children }: { children: ReactNode }) => {
     if (selectedTag) {
       fetchPostsByTag(selectedTag);
     } else {
-      fetchPosts();
+      fetchSetPosts();
     }
     updateURL();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [skip, limit, sortBy, sortOrder, selectedTag]); // TODO 수정예정
-
-  useEffect(() => {
-    const params = new URLSearchParams(location.search);
-    setSkip(parseInt(params.get("skip") || "0"));
-    setLimit(parseInt(params.get("limit") || "10"));
-    setSearchQuery(params.get("search") || "");
-    setSortBy(params.get("sortBy") || "");
-    setSortOrder(params.get("sortOrder") || "asc");
-    setSelectedTag(params.get("tag") || "");
-  }, [location.search]);
 
   const value = {
     posts,
@@ -168,26 +93,12 @@ export const PostProvider = ({ children }: { children: ReactNode }) => {
     setSelectedPost,
     total,
     setTotal,
-    skip,
-    setSkip,
-    limit,
-    setLimit,
-    searchQuery,
-    setSearchQuery,
-    sortBy,
-    setSortBy,
-    sortOrder,
-    setSortOrder,
     tags,
     setTags,
-    selectedTag,
-    setSelectedTag,
     loading,
     setLoading,
-    fetchPosts,
+    fetchSetPosts,
     fetchPostsByTag,
-    updateURL,
-    deletePost,
   };
 
   return <PostContext.Provider value={value}>{children}</PostContext.Provider>;
