+++ src/app/providers/index.tsx
@@ -2,17 +2,20 @@ import { ReactNode } from "react";
 import { CommentProvider } from "./CommentProvider";
 import { DialogProvider } from "./DialogProvider";
 import { PostProvider } from "./PostProvider";
+import { URLProvider } from "./URLProvider";
 import { UserProvider } from "./UserProvider";
 
 const Providers = ({ children }: { children: ReactNode }) => {
   return (
-    <PostProvider>
-      <CommentProvider>
-        <UserProvider>
-          <DialogProvider>{children}</DialogProvider>
-        </UserProvider>
-      </CommentProvider>
-    </PostProvider>
+    <URLProvider>
+      <PostProvider>
+        <CommentProvider>
+          <UserProvider>
+            <DialogProvider>{children}</DialogProvider>
+          </UserProvider>
+        </CommentProvider>
+      </PostProvider>
+    </URLProvider>
   );
 };
